+++ docs/release_notes.md
@@ -6,7 +6,7 @@
 
 1. Changing `get/interfaces` `lag` to be a parent lag name and adding `lag_id` attribute
 2. `get/interfaces` for Arista adding interfaces speeds map
-
+3. Enhancing `get/interfaces` for Arista to calculate access VLAN id for SVI interfaces
 ---
 
 ## 0.5.5